+++ runtime/vm/code_descriptors.cc
@@ -77,7 +77,7 @@ RawArray* StackmapTableBuilder::FinalizeStackmaps(const Code& code) {
 }
 
 
-RawStackmap* StackmapTableBuilder::MapAt(int index) const {
+RawStackmap* StackmapTableBuilder::MapAt(intptr_t index) const {
   Stackmap& map = Stackmap::Handle();
   map ^= list_.At(index);
   return map.raw();
